Add CoreWingCrossSectionMovement to _core.py
Extract shared WingCrossSectionMovement infrastructure into
CoreWingCrossSectionMovement in _core.py, following the same pattern
used for CoreOperatingPointMovement. Refactor WingCrossSectionMovement
to extend it with an empty __slots__ and a pass-through __init__.
