+++ pterasoftware/_core.py
@@ -2,10 +2,13 @@
 
 from __future__ import annotations
 
+import copy
 import math
-from collections.abc import Callable
+from collections.abc import Callable, Sequence
 
-from . import _parameter_validation
+import numpy as np
+
+from . import _parameter_validation, geometry
 from . import operating_point as operating_point_mod
 from .movements import _functions
 from .movements import airplane_movement as airplane_movement_mod
@@ -214,6 +217,481 @@ class CoreOperatingPointMovement:
             CgP1_E_Eo=self._base_operating_point.CgP1_E_Eo,
             surfaceNormal_E=self._base_operating_point.surfaceNormal_E,
             surfacePoint_E_Eo=self._base_operating_point.surfacePoint_E_Eo,
+        )
+
+
+class CoreWingCrossSectionMovement:
+    """A core class used to contain the shared foundation of WingCrossSectionMovement
+    and its feature variant siblings.
+
+    See WingCrossSectionMovement for full documentation of the shared interface.
+
+    CoreWingCrossSectionMovement holds the base WingCrossSection and oscillation
+    parameters, and provides generate_wing_cross_section_at_time_step() for creating
+    WingCrossSections one step at a time.
+    """
+
+    __slots__ = (
+        "_base_wing_cross_section",
+        "_ampLp_Wcsp_Lpp",
+        "_periodLp_Wcsp_Lpp",
+        "_spacingLp_Wcsp_Lpp",
+        "_phaseLp_Wcsp_Lpp",
+        "_ampAngles_Wcsp_to_Wcs_ixyz",
+        "_periodAngles_Wcsp_to_Wcs_ixyz",
+        "_spacingAngles_Wcsp_to_Wcs_ixyz",
+        "_phaseAngles_Wcsp_to_Wcs_ixyz",
+        "_all_periods",
+        "_max_period",
+    )
+
+    def __init__(
+        self,
+        base_wing_cross_section: geometry.wing_cross_section.WingCrossSection,
+        ampLp_Wcsp_Lpp: np.ndarray | Sequence[float | int] = (0.0, 0.0, 0.0),
+        periodLp_Wcsp_Lpp: np.ndarray | Sequence[float | int] = (0.0, 0.0, 0.0),
+        spacingLp_Wcsp_Lpp: np.ndarray | Sequence[str | Callable[[float], float]] = (
+            "sine",
+            "sine",
+            "sine",
+        ),
+        phaseLp_Wcsp_Lpp: np.ndarray | Sequence[float | int] = (0.0, 0.0, 0.0),
+        ampAngles_Wcsp_to_Wcs_ixyz: np.ndarray | Sequence[float | int] = (
+            0.0,
+            0.0,
+            0.0,
+        ),
+        periodAngles_Wcsp_to_Wcs_ixyz: np.ndarray | Sequence[float | int] = (
+            0.0,
+            0.0,
+            0.0,
+        ),
+        spacingAngles_Wcsp_to_Wcs_ixyz: (
+            np.ndarray | Sequence[str | Callable[[float], float]]
+        ) = (
+            "sine",
+            "sine",
+            "sine",
+        ),
+        phaseAngles_Wcsp_to_Wcs_ixyz: np.ndarray | Sequence[float | int] = (
+            0.0,
+            0.0,
+            0.0,
+        ),
+    ) -> None:
+        """The initialization method.
+
+        See WingCrossSectionMovement's initialization method for full parameter
+        descriptions.
+
+        :param base_wing_cross_section: The base WingCrossSection.
+        :param ampLp_Wcsp_Lpp: The amplitudes of Lp_Wcsp_Lpp oscillation in meters.
+        :param periodLp_Wcsp_Lpp: The periods of Lp_Wcsp_Lpp oscillation in seconds.
+        :param spacingLp_Wcsp_Lpp: The spacing types for Lp_Wcsp_Lpp oscillation.
+        :param phaseLp_Wcsp_Lpp: The phase offsets of Lp_Wcsp_Lpp oscillation in
+            degrees.
+        :param ampAngles_Wcsp_to_Wcs_ixyz: The amplitudes of angles_Wcsp_to_Wcs_ixyz
+            oscillation in degrees.
+        :param periodAngles_Wcsp_to_Wcs_ixyz: The periods of angles_Wcsp_to_Wcs_ixyz
+            oscillation in seconds.
+        :param spacingAngles_Wcsp_to_Wcs_ixyz: The spacing types for
+            angles_Wcsp_to_Wcs_ixyz oscillation.
+        :param phaseAngles_Wcsp_to_Wcs_ixyz: The phase offsets of
+            angles_Wcsp_to_Wcs_ixyz oscillation in degrees.
+        :return: None
+        """
+        # Validate and store immutable attributes. Set those that are numpy
+        # arrays to be read only.
+        if not isinstance(
+            base_wing_cross_section,
+            geometry.wing_cross_section.WingCrossSection,
+        ):
+            raise TypeError("base_wing_cross_section must be a WingCrossSection.")
+        self._base_wing_cross_section = base_wing_cross_section
+
+        ampLp_Wcsp_Lpp = _parameter_validation.threeD_number_vectorLike_return_float(
+            ampLp_Wcsp_Lpp, "ampLp_Wcsp_Lpp"
+        )
+        if not np.all(ampLp_Wcsp_Lpp >= 0.0):
+            raise ValueError("All elements in ampLp_Wcsp_Lpp must be non negative.")
+        self._ampLp_Wcsp_Lpp = ampLp_Wcsp_Lpp
+        self._ampLp_Wcsp_Lpp.flags.writeable = False
+
+        periodLp_Wcsp_Lpp = _parameter_validation.threeD_number_vectorLike_return_float(
+            periodLp_Wcsp_Lpp, "periodLp_Wcsp_Lpp"
+        )
+        if not np.all(periodLp_Wcsp_Lpp >= 0.0):
+            raise ValueError("All elements in periodLp_Wcsp_Lpp must be non negative.")
+        for period_index, period in enumerate(periodLp_Wcsp_Lpp):
+            amp = self._ampLp_Wcsp_Lpp[period_index]
+            if amp == 0 and period != 0:
+                raise ValueError(
+                    "If an element in ampLp_Wcsp_Lpp is 0.0, the "
+                    "corresponding element in periodLp_Wcsp_Lpp must be "
+                    "also be 0.0."
+                )
+        self._periodLp_Wcsp_Lpp = periodLp_Wcsp_Lpp
+        self._periodLp_Wcsp_Lpp.flags.writeable = False
+
+        # Store as tuple to prevent external mutation.
+        self._spacingLp_Wcsp_Lpp = (
+            _parameter_validation.threeD_spacing_vectorLike_return_tuple(
+                spacingLp_Wcsp_Lpp, "spacingLp_Wcsp_Lpp"
+            )
+        )
+
+        phaseLp_Wcsp_Lpp = _parameter_validation.threeD_number_vectorLike_return_float(
+            phaseLp_Wcsp_Lpp, "phaseLp_Wcsp_Lpp"
+        )
+        if not (
+            np.all(phaseLp_Wcsp_Lpp > -180.0) and np.all(phaseLp_Wcsp_Lpp <= 180.0)
+        ):
+            raise ValueError(
+                "All elements in phaseLp_Wcsp_Lpp must be in the range "
+                "(-180.0, 180.0]."
+            )
+        for phase_index, phase in enumerate(phaseLp_Wcsp_Lpp):
+            amp = self._ampLp_Wcsp_Lpp[phase_index]
+            if amp == 0 and phase != 0:
+                raise ValueError(
+                    "If an element in ampLp_Wcsp_Lpp is 0.0, the "
+                    "corresponding element in phaseLp_Wcsp_Lpp must be "
+                    "also be 0.0."
+                )
+        self._phaseLp_Wcsp_Lpp = phaseLp_Wcsp_Lpp
+        self._phaseLp_Wcsp_Lpp.flags.writeable = False
+
+        ampAngles_Wcsp_to_Wcs_ixyz = (
+            _parameter_validation.threeD_number_vectorLike_return_float(
+                ampAngles_Wcsp_to_Wcs_ixyz, "ampAngles_Wcsp_to_Wcs_ixyz"
+            )
+        )
+        if not (
+            np.all(ampAngles_Wcsp_to_Wcs_ixyz >= 0.0)
+            and np.all(ampAngles_Wcsp_to_Wcs_ixyz <= 180.0)
+        ):
+            raise ValueError(
+                "All elements in ampAngles_Wcsp_to_Wcs_ixyz must be in "
+                "the range [0.0, 180.0]."
+            )
+        self._ampAngles_Wcsp_to_Wcs_ixyz = ampAngles_Wcsp_to_Wcs_ixyz
+        self._ampAngles_Wcsp_to_Wcs_ixyz.flags.writeable = False
+
+        periodAngles_Wcsp_to_Wcs_ixyz = (
+            _parameter_validation.threeD_number_vectorLike_return_float(
+                periodAngles_Wcsp_to_Wcs_ixyz,
+                "periodAngles_Wcsp_to_Wcs_ixyz",
+            )
+        )
+        if not np.all(periodAngles_Wcsp_to_Wcs_ixyz >= 0.0):
+            raise ValueError(
+                "All elements in periodAngles_Wcsp_to_Wcs_ixyz must be " "non negative."
+            )
+        for period_index, period in enumerate(periodAngles_Wcsp_to_Wcs_ixyz):
+            amp = self._ampAngles_Wcsp_to_Wcs_ixyz[period_index]
+            if amp == 0 and period != 0:
+                raise ValueError(
+                    "If an element in ampAngles_Wcsp_to_Wcs_ixyz is 0.0, "
+                    "the corresponding element in "
+                    "periodAngles_Wcsp_to_Wcs_ixyz must be also be 0.0."
+                )
+        self._periodAngles_Wcsp_to_Wcs_ixyz = periodAngles_Wcsp_to_Wcs_ixyz
+        self._periodAngles_Wcsp_to_Wcs_ixyz.flags.writeable = False
+
+        # Store as tuple to prevent external mutation.
+        self._spacingAngles_Wcsp_to_Wcs_ixyz = (
+            _parameter_validation.threeD_spacing_vectorLike_return_tuple(
+                spacingAngles_Wcsp_to_Wcs_ixyz,
+                "spacingAngles_Wcsp_to_Wcs_ixyz",
+            )
+        )
+
+        phaseAngles_Wcsp_to_Wcs_ixyz = (
+            _parameter_validation.threeD_number_vectorLike_return_float(
+                phaseAngles_Wcsp_to_Wcs_ixyz,
+                "phaseAngles_Wcsp_to_Wcs_ixyz",
+            )
+        )
+        if not (
+            np.all(phaseAngles_Wcsp_to_Wcs_ixyz > -180.0)
+            and np.all(phaseAngles_Wcsp_to_Wcs_ixyz <= 180.0)
+        ):
+            raise ValueError(
+                "All elements in phaseAngles_Wcsp_to_Wcs_ixyz must be in "
+                "the range (-180.0, 180.0]."
+            )
+        for phase_index, phase in enumerate(phaseAngles_Wcsp_to_Wcs_ixyz):
+            amp = self._ampAngles_Wcsp_to_Wcs_ixyz[phase_index]
+            if amp == 0 and phase != 0:
+                raise ValueError(
+                    "If an element in ampAngles_Wcsp_to_Wcs_ixyz is 0.0, "
+                    "the corresponding element in "
+                    "phaseAngles_Wcsp_to_Wcs_ixyz must be also be 0.0."
+                )
+        self._phaseAngles_Wcsp_to_Wcs_ixyz = phaseAngles_Wcsp_to_Wcs_ixyz
+        self._phaseAngles_Wcsp_to_Wcs_ixyz.flags.writeable = False
+
+        # Initialize the caches for the properties derived from the immutable
+        # attributes.
+        self._all_periods: tuple[float, ...] | None = None
+        self._max_period: float | None = None
+
+    # --- Deep copy method ---
+    def __deepcopy__(self, memo: dict) -> CoreWingCrossSectionMovement:
+        """Creates a deep copy of this CoreWingCrossSectionMovement.
+
+        See WingCrossSectionMovement for full documentation.
+
+        :param memo: A dict used by the copy module to track already copied objects and
+            avoid infinite recursion.
+        :return: A new instance with copied attributes.
+        """
+        # Create a new instance without calling __init__ to avoid redundant
+        # validation. Use type(self) so subclasses get the correct type.
+        new_movement = object.__new__(type(self))
+
+        # Store this instance in memo to handle potential circular references.
+        memo[id(self)] = new_movement
+
+        # Deep copy the base WingCrossSection to ensure independence
+        # (immutable).
+        new_movement._base_wing_cross_section = copy.deepcopy(
+            self._base_wing_cross_section, memo
+        )
+
+        # Copy numpy arrays and make them read only.
+        new_movement._ampLp_Wcsp_Lpp = self._ampLp_Wcsp_Lpp.copy()
+        new_movement._ampLp_Wcsp_Lpp.flags.writeable = False
+
+        new_movement._periodLp_Wcsp_Lpp = self._periodLp_Wcsp_Lpp.copy()
+        new_movement._periodLp_Wcsp_Lpp.flags.writeable = False
+
+        new_movement._phaseLp_Wcsp_Lpp = self._phaseLp_Wcsp_Lpp.copy()
+        new_movement._phaseLp_Wcsp_Lpp.flags.writeable = False
+
+        new_movement._ampAngles_Wcsp_to_Wcs_ixyz = (
+            self._ampAngles_Wcsp_to_Wcs_ixyz.copy()
+        )
+        new_movement._ampAngles_Wcsp_to_Wcs_ixyz.flags.writeable = False
+
+        new_movement._periodAngles_Wcsp_to_Wcs_ixyz = (
+            self._periodAngles_Wcsp_to_Wcs_ixyz.copy()
+        )
+        new_movement._periodAngles_Wcsp_to_Wcs_ixyz.flags.writeable = False
+
+        new_movement._phaseAngles_Wcsp_to_Wcs_ixyz = (
+            self._phaseAngles_Wcsp_to_Wcs_ixyz.copy()
+        )
+        new_movement._phaseAngles_Wcsp_to_Wcs_ixyz.flags.writeable = False
+
+        # Copy tuples directly (they are immutable).
+        new_movement._spacingLp_Wcsp_Lpp = self._spacingLp_Wcsp_Lpp
+        new_movement._spacingAngles_Wcsp_to_Wcs_ixyz = (
+            self._spacingAngles_Wcsp_to_Wcs_ixyz
+        )
+
+        # Initialize cache variables to None (caches will be recomputed on
+        # access).
+        new_movement._all_periods = None
+        new_movement._max_period = None
+
+        return new_movement
+
+    # --- Immutable: read only properties ---
+    @property
+    def base_wing_cross_section(
+        self,
+    ) -> geometry.wing_cross_section.WingCrossSection:
+        return self._base_wing_cross_section
+
+    @property
+    def ampLp_Wcsp_Lpp(self) -> np.ndarray:
+        return self._ampLp_Wcsp_Lpp
+
+    @property
+    def periodLp_Wcsp_Lpp(self) -> np.ndarray:
+        return self._periodLp_Wcsp_Lpp
+
+    @property
+    def spacingLp_Wcsp_Lpp(
+        self,
+    ) -> tuple[str | Callable[[float], float], ...]:
+        return self._spacingLp_Wcsp_Lpp
+
+    @property
+    def phaseLp_Wcsp_Lpp(self) -> np.ndarray:
+        return self._phaseLp_Wcsp_Lpp
+
+    @property
+    def ampAngles_Wcsp_to_Wcs_ixyz(self) -> np.ndarray:
+        return self._ampAngles_Wcsp_to_Wcs_ixyz
+
+    @property
+    def periodAngles_Wcsp_to_Wcs_ixyz(self) -> np.ndarray:
+        return self._periodAngles_Wcsp_to_Wcs_ixyz
+
+    @property
+    def spacingAngles_Wcsp_to_Wcs_ixyz(
+        self,
+    ) -> tuple[str | Callable[[float], float], ...]:
+        return self._spacingAngles_Wcsp_to_Wcs_ixyz
+
+    @property
+    def phaseAngles_Wcsp_to_Wcs_ixyz(self) -> np.ndarray:
+        return self._phaseAngles_Wcsp_to_Wcs_ixyz
+
+    # --- Immutable derived: manual lazy caching ---
+    @property
+    def all_periods(self) -> tuple[float, ...]:
+        """All unique non zero periods from this CoreWingCrossSectionMovement.
+
+        See WingCrossSectionMovement for full documentation.
+
+        :return: A tuple of all unique non zero periods in seconds. If the motion is
+            static, this will be an empty tuple.
+        """
+        if self._all_periods is None:
+            periods = []
+
+            # Collect all periods from positional motion.
+            for period in self._periodLp_Wcsp_Lpp:
+                if period > 0.0:
+                    periods.append(float(period))
+
+            # Collect all periods from angular motion.
+            for period in self._periodAngles_Wcsp_to_Wcs_ixyz:
+                if period > 0.0:
+                    periods.append(float(period))
+
+            self._all_periods = tuple(periods)
+        return self._all_periods
+
+    @property
+    def max_period(self) -> float:
+        """CoreWingCrossSectionMovement's longest period of motion.
+
+        See WingCrossSectionMovement for full documentation.
+
+        :return: The longest period in seconds. If the motion is static, this will be
+            0.0.
+        """
+        if self._max_period is None:
+            self._max_period = float(
+                max(
+                    np.max(self._periodLp_Wcsp_Lpp),
+                    np.max(self._periodAngles_Wcsp_to_Wcs_ixyz),
+                )
+            )
+        return self._max_period
+
+    # --- Other methods ---
+    def generate_wing_cross_section_at_time_step(
+        self, step: int, delta_time: float | int
+    ) -> geometry.wing_cross_section.WingCrossSection:
+        """Creates the WingCrossSection at a single time step.
+
+        See WingCrossSectionMovement for full documentation.
+
+        :param step: The time step index. Must be a non negative int.
+        :param delta_time: The time between each time step in seconds. Must be a
+            positive number (int or float).
+        :return: The WingCrossSection at this time step.
+        """
+        time = step * delta_time
+
+        # Evaluate the oscillating value for each dimension of Lp_Wcsp_Lpp.
+        thisLp_Wcsp_Lpp = np.zeros(3, dtype=float)
+        for dim in range(3):
+            this_spacing = self._spacingLp_Wcsp_Lpp[dim]
+            this_amp = self._ampLp_Wcsp_Lpp[dim]
+            this_period = self._periodLp_Wcsp_Lpp[dim]
+            this_phase = self._phaseLp_Wcsp_Lpp[dim]
+            this_base = self._base_wing_cross_section.Lp_Wcsp_Lpp[dim]
+
+            if this_spacing == "sine":
+                thisLp_Wcsp_Lpp[dim] = _functions.oscillating_sin_at_time(
+                    amp=this_amp,
+                    period=this_period,
+                    phase=this_phase,
+                    base=this_base,
+                    time=time,
+                )
+            elif this_spacing == "uniform":
+                thisLp_Wcsp_Lpp[dim] = _functions.oscillating_lin_at_time(
+                    amp=this_amp,
+                    period=this_period,
+                    phase=this_phase,
+                    base=this_base,
+                    time=time,
+                )
+            elif callable(this_spacing):
+                thisLp_Wcsp_Lpp[dim] = _functions.oscillating_custom_at_time(
+                    amp=this_amp,
+                    period=this_period,
+                    phase=this_phase,
+                    base=this_base,
+                    time=time,
+                    custom_function=this_spacing,
+                )
+            else:
+                raise ValueError(f"Invalid spacing value: {this_spacing}")
+
+        # Evaluate the oscillating value for each dimension of
+        # angles_Wcsp_to_Wcs_ixyz.
+        theseAngles_Wcsp_to_Wcs_ixyz = np.zeros(3, dtype=float)
+        for dim in range(3):
+            this_spacing = self._spacingAngles_Wcsp_to_Wcs_ixyz[dim]
+            this_amp = self._ampAngles_Wcsp_to_Wcs_ixyz[dim]
+            this_period = self._periodAngles_Wcsp_to_Wcs_ixyz[dim]
+            this_phase = self._phaseAngles_Wcsp_to_Wcs_ixyz[dim]
+            this_base = self._base_wing_cross_section.angles_Wcsp_to_Wcs_ixyz[dim]
+
+            if this_spacing == "sine":
+                theseAngles_Wcsp_to_Wcs_ixyz[dim] = _functions.oscillating_sin_at_time(
+                    amp=this_amp,
+                    period=this_period,
+                    phase=this_phase,
+                    base=this_base,
+                    time=time,
+                )
+            elif this_spacing == "uniform":
+                theseAngles_Wcsp_to_Wcs_ixyz[dim] = _functions.oscillating_lin_at_time(
+                    amp=this_amp,
+                    period=this_period,
+                    phase=this_phase,
+                    base=this_base,
+                    time=time,
+                )
+            elif callable(this_spacing):
+                theseAngles_Wcsp_to_Wcs_ixyz[dim] = (
+                    _functions.oscillating_custom_at_time(
+                        amp=this_amp,
+                        period=this_period,
+                        phase=this_phase,
+                        base=this_base,
+                        time=time,
+                        custom_function=this_spacing,
+                    )
+                )
+            else:
+                raise ValueError(f"Invalid spacing value: {this_spacing}")
+
+        return geometry.wing_cross_section.WingCrossSection(
+            airfoil=self._base_wing_cross_section.airfoil,
+            num_spanwise_panels=(self._base_wing_cross_section.num_spanwise_panels),
+            chord=self._base_wing_cross_section.chord,
+            Lp_Wcsp_Lpp=thisLp_Wcsp_Lpp,
+            angles_Wcsp_to_Wcs_ixyz=theseAngles_Wcsp_to_Wcs_ixyz,
+            control_surface_symmetry_type=(
+                self._base_wing_cross_section.control_surface_symmetry_type
+            ),
+            control_surface_hinge_point=(
+                self._base_wing_cross_section.control_surface_hinge_point
+            ),
+            control_surface_deflection=(
+                self._base_wing_cross_section.control_surface_deflection
+            ),
+            spanwise_spacing=(self._base_wing_cross_section.spanwise_spacing),
         )
